Added method to signer class
